Unlinked activity, revisit
Split the student and instructor stuff into two separate components.
Reset the student workflow to function correctly and improve readability of instructions.
Reset the instructor workflow too. But will need continue work on listing of learning objects and testing create new content.  
Need to replicate what the learning objects page does for pagination.
Need to be sure the create also links.
